LayoutTests/platform/mac-monterey/http/tests/security/xslt-external-entity-expected.txt

CONSOLE MESSAGE: Failure to process entity passwd

CONSOLE MESSAGE: Entity 'passwd' not defined

CONSOLE MESSAGE: Failure to process entity hosts

CONSOLE MESSAGE: Entity 'hosts' not defined

CONSOLE MESSAGE: Failure to process entity group

CONSOLE MESSAGE: Entity 'group' not defined

ALERT: document()
